compositor|tests: Split the sample tests to follow their subjects
Assertions about the representative sample's artifacts and files now sit beside the builders that produce them. The shared sample tests keep the properties every published sample satisfies.
